Add CPU support
