+++ MaxText/max_utils.py
@@ -169,13 +169,13 @@ def upload_blob(destination_gcs_name, source_file_name):
   blob = bucket.blob(prefix_name)
   blob.upload_from_filename(source_file_name)
 
-def upload_csv(csv_file, data, gcs_folder):
-  """Upload csv file for each process to GCS"""
+def upload_csv(csv_file, data, gcs_path):
+  """Upload the csv files generated by each process to GCS"""
   with open(csv_file, 'w', encoding="utf-8", newline='') as file:
-    w = csv.writer(file)
+    csv_writer = csv.writer(file)
     for _, t in enumerate(data):
-      w.writerow(t)
-  upload_blob(f"{gcs_folder}/{csv_file}", csv_file)
+      csv_writer.writerow(t)
+  upload_blob(gcs_path, csv_file)
 
 def download_blob(source_gcs_name, destination_file_name):
   """Downloads a file from a GCS location and save to a local file"""
